+++ diive/pkgs/outlierdetection/__init__.py
@@ -1,3 +1,4 @@
 from . import absolutelimits
 from . import thymeboost
 from . import missing
+